admin/backend/pyproject.toml
[project]
name = "admin-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "opensearch-py>=3.1.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.2.2",
    "lightrag-hku>=1.4.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../../tests/admin"]
